+++ backend/src/kitconcept/voltolighttheme/__init__.py
@@ -5,7 +5,7 @@ from zope.i18nmessageid import MessageFactory
 import logging
 
 
-__version__ = "7.0.0a8"
+__version__ = "7.0.0a12"
 
 PACKAGE_NAME = "kitconcept.voltolighttheme"