+++ frontend/CHANGELOG.md
@@ -8,6 +8,48 @@
 
 <!-- towncrier release notes start -->
 
+## 7.0.0-alpha.12 (2025-06-18)
+
+### Feature
+
+- Reintroduce an improved `Anontools` using `portal_actions`. @sneridagh [#581](https://github.com/kitconcept/volto-light-theme/pull/581)
+
+### Bugfix
+
+- Fix sticky menu precedence in z-index with the fat menu. @sneridagh [#582](https://github.com/kitconcept/volto-light-theme/pull/582)
+- Fixed listing in grids variations to match the standalone ones (except Grid). @sneridagh [#582](https://github.com/kitconcept/volto-light-theme/pull/582)
+- Fixed Person CSS in Search block. @sneridagh [#584](https://github.com/kitconcept/volto-light-theme/pull/584)
+
+## 7.0.0-alpha.11 (2025-06-12)
+
+### Bugfix
+
+- Do not show sticky menu in add view. @sneridagh 
+- Fix aspect ratio for images in file grid teaser block @iRohitSingh 
+- Fixed position of sticky menu for small screens. @sneridagh 
+
+### Internal
+
+- Added visual regression tests for Storybook stories. @sneridagh [#576](https://github.com/kitconcept/volto-light-theme/pull/576)
+- Upgrade to Volto 18.23.0 @sneridagh 
+- Use `vitest` for unit tests. @sneridagh 
+
+## 7.0.0-alpha.10 (2025-06-10)
+
+### Bugfix
+
+- Fixed Person Teaser top variant. Added Storybook. @sneridagh 
+
+## 7.0.0-alpha.9 (2025-06-09)
+
+### Bugfix
+
+- Fixed Teaser styling @sneridagh [#573](https://github.com/kitconcept/volto-light-theme/pull/573)
+
+### Internal
+
+- Fix aspect ratio for images in file teaser block @iRohitSingh 
+
 ## 7.0.0-alpha.8 (2025-06-06)
 
 ### Bugfix